+++ packages/xchain-thorchain-query/CHANGELOG.md
@@ -1,3 +1,14 @@
+# v0.2.6 (2023-4-24)
+
+## Add
+
+- Bump packages
+# v0.2.5 (2023-4-14)
+
+## Fix
+
+- Savers Error handling
+
 # v0.2.4 (2023-4-05)
 
 ## Fix